V3 ports (#438)
* feat(agent): add timeout on write calls (#426)

* feat(agent): add timeout on write calls

When sending a message via agent, the write connection does not have
a default timeout, leading to blocking calls if the connections is
unhealthy or simply a bad route to it. Thus, adding a write timeout
to the writes so we fail but don't block. If deadline is exceeded then
we log an error but don't close a connection, since it might fail due
to bad route but not a broken pipe per-se. The new configuration can
be found under: `pitaya.buffer.agent.writetimeout`, with 10s by deafult.

The timeout is set using `SetWriteDeadline` function from go net pkg:
https://pkg.go.dev/net#Conn

* fix(agent): wrap panic error in broken pipe

* debug(agent): add error messages on send failures

* Nats connection options (#437)

* Replace deprecated DisconnectHandler with DisconnectErrHandler and provide output of the underlying disconnection error

* Add additional nats configuration options

* Bump nats client from 1.25.0 to 1.38.0

* Update docs with new nats options

* fix duplicate config option

* remove initial connection retry for now

---------

Co-authored-by: Pedro Soares <pedro.soares@wildlifestudios.com>